Refactor: move nestedXPathField into src/widgets/
