+++ src/widgets/cardList.js
@@ -1,7 +1,7 @@
 import _ from "underscore";
 import $ from "jquery";
 import atwho from "vellum/atwho";
-import nestedXPathField from "vellum/nestedXPathField";
+import nestedXPathField from "vellum/widgets/nestedXPathField";
 import {normal} from "vellum/widgets/base";
 import widget_card_list from "vellum/templates/widget_card_list.html";
 import nested_xpath_field from "vellum/templates/nested_xpath_field.html";
