+++ src/widgets/nestedXPathField.js
